requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.2
    # via requests
cython==3.0.5
    # via python-frog
folia==2.5.8
    # via tscan (setup.py)
idna==3.4
    # via requests
isodate==0.6.1
    # via rdflib
lxml==4.9.3
    # via folia
pyparsing==3.1.1
    # via rdflib
python-frog==0.6.8
    # via tscan (setup.py)
rdflib==7.0.0
    # via folia
requests==2.31.0
    # via folia
six==1.16.0
    # via isodate
urllib3==2.0.7
    # via requests
